+++ sdk/devhub/arm-devhub/src/models/index.ts
@@ -242,6 +242,30 @@ export interface WorkflowRun {
   readonly lastRunAt?: Date;
 }
 
+/** Properties used for generating artifacts such as Dockerfiles and manifests. */
+export interface ArtifactGenerationProperties {
+  /** The programming language used. */
+  generationLanguage?: GenerationLanguage;
+  /** The version of the language image used for execution in the generated dockerfile. */
+  languageVersion?: string;
+  /** The version of the language image used for building the code in the generated dockerfile. */
+  builderVersion?: string;
+  /** The port the application is exposed on. */
+  port?: string;
+  /** The name of the app. */
+  appName?: string;
+  /** Determines the type of manifests within the repository. */
+  manifestType?: ManifestType;
+  /** The mode of generation to be used for generating Dockerfiles. */
+  dockerfileGenerationMode?: DockerfileGenerationMode;
+  /** The mode of generation to be used for generating Dockerfiles. */
+  manifestGenerationMode?: ManifestGenerationMode;
+  /** The directory to output the generated Dockerfile to. */
+  dockerfileOutputDirectory?: string;
+  /** The directory to output the generated manifests to. */
+  manifestOutputDirectory?: string;
+}
+
 /** delete response if content must be provided on delete operation */
 export interface DeleteWorkflowResponse {
   /** delete status message */
@@ -308,8 +332,13 @@ export interface Workflow extends TrackedResource {
    */
   readonly prStatus?: PullRequestStatus;
   lastWorkflowRun?: WorkflowRun;
-  /** Determines the type of manifests within the repository. */
-  authStatus?: ManifestType;
+  /**
+   * Determines the authorization status of requests.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly authStatus?: AuthorizationStatus;
+  /** Properties used for generating artifacts such as Dockerfiles and manifests. */
+  artifactGenerationProperties?: ArtifactGenerationProperties;
 }
 
 /** Known values of {@link Origin} that the service accepts. */
@@ -374,10 +403,12 @@ export type CreatedByType = string;
 
 /** Known values of {@link ManifestType} that the service accepts. */
 export enum KnownManifestType {
-  /** Repositories using helm */
+  /** Manifests using helm */
   Helm = "helm",
-  /** Repositories using kubernetes manifests */
-  Kube = "kube"
+  /** Standard Kubernetes manifests */
+  Kube = "kube",
+  /** Manifests using kustomize */
+  Kustomize = "kustomize"
 }
 
 /**
@@ -385,8 +416,9 @@ export enum KnownManifestType {
  * {@link KnownManifestType} can be used interchangeably with ManifestType,
  *  this enum contains the known values that the service supports.
  * ### Known values supported by the service
- * **helm**: Repositories using helm \
- * **kube**: Repositories using kubernetes manifests
+ * **helm**: Manifests using helm \
+ * **kube**: Standard Kubernetes manifests \
+ * **kustomize**: Manifests using kustomize
  */
 export type ManifestType = string;
 
@@ -413,6 +445,114 @@ export enum KnownPullRequestStatus {
  * **removed**: Workflow no longer found within repository.
  */
 export type PullRequestStatus = string;
+
+/** Known values of {@link AuthorizationStatus} that the service accepts. */
+export enum KnownAuthorizationStatus {
+  /** Requests authorized successfully */
+  Authorized = "Authorized",
+  /** Requests returned NotFound response */
+  NotFound = "NotFound",
+  /** Requests returned other error response */
+  Error = "Error"
+}
+
+/**
+ * Defines values for AuthorizationStatus. \
+ * {@link KnownAuthorizationStatus} can be used interchangeably with AuthorizationStatus,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Authorized**: Requests authorized successfully \
+ * **NotFound**: Requests returned NotFound response \
+ * **Error**: Requests returned other error response
+ */
+export type AuthorizationStatus = string;
+
+/** Known values of {@link GenerationLanguage} that the service accepts. */
+export enum KnownGenerationLanguage {
+  /** clojure language */
+  Clojure = "clojure",
+  /** csharp language */
+  Csharp = "csharp",
+  /** erlang language */
+  Erlang = "erlang",
+  /** go language */
+  Go = "go",
+  /** gomodule language */
+  Gomodule = "gomodule",
+  /** gradle language */
+  Gradle = "gradle",
+  /** java language */
+  Java = "java",
+  /** javascript language */
+  Javascript = "javascript",
+  /** php language */
+  Php = "php",
+  /** python language */
+  Python = "python",
+  /** ruby language */
+  Ruby = "ruby",
+  /** rust language */
+  Rust = "rust",
+  /** swift language */
+  Swift = "swift"
+}
+
+/**
+ * Defines values for GenerationLanguage. \
+ * {@link KnownGenerationLanguage} can be used interchangeably with GenerationLanguage,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **clojure**: clojure language \
+ * **csharp**: csharp language \
+ * **erlang**: erlang language \
+ * **go**: go language \
+ * **gomodule**: gomodule language \
+ * **gradle**: gradle language \
+ * **java**: java language \
+ * **javascript**: javascript language \
+ * **php**: php language \
+ * **python**: python language \
+ * **ruby**: ruby language \
+ * **rust**: rust language \
+ * **swift**: swift language
+ */
+export type GenerationLanguage = string;
+
+/** Known values of {@link DockerfileGenerationMode} that the service accepts. */
+export enum KnownDockerfileGenerationMode {
+  /** Dockerfiles will be generated */
+  Enabled = "enabled",
+  /** Dockerfiles will not be generated */
+  Disabled = "disabled"
+}
+
+/**
+ * Defines values for DockerfileGenerationMode. \
+ * {@link KnownDockerfileGenerationMode} can be used interchangeably with DockerfileGenerationMode,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **enabled**: Dockerfiles will be generated \
+ * **disabled**: Dockerfiles will not be generated
+ */
+export type DockerfileGenerationMode = string;
+
+/** Known values of {@link ManifestGenerationMode} that the service accepts. */
+export enum KnownManifestGenerationMode {
+  /** Manifests will be generated */
+  Enabled = "enabled",
+  /** Manifests will not be generated */
+  Disabled = "disabled"
+}
+
+/**
+ * Defines values for ManifestGenerationMode. \
+ * {@link KnownManifestGenerationMode} can be used interchangeably with ManifestGenerationMode,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **enabled**: Manifests will be generated \
+ * **disabled**: Manifests will not be generated
+ */
+export type ManifestGenerationMode = string;
 
 /** Optional parameters. */
 export interface OperationsListOptionalParams
@@ -453,10 +593,7 @@ export type WorkflowListResponse = WorkflowListResult;
 
 /** Optional parameters. */
 export interface WorkflowListByResourceGroupOptionalParams
-  extends coreClient.OperationOptions {
-  /** The ManagedCluster resource associated with the workflows. */
-  managedClusterResource?: string;
-}
+  extends coreClient.OperationOptions {}
 
 /** Contains response data for the listByResourceGroup operation. */
 export type WorkflowListByResourceGroupResponse = WorkflowListResult;
@@ -498,10 +635,7 @@ export type WorkflowListNextResponse = WorkflowListResult;
 
 /** Optional parameters. */
 export interface WorkflowListByResourceGroupNextOptionalParams
-  extends coreClient.OperationOptions {
-  /** The ManagedCluster resource associated with the workflows. */
-  managedClusterResource?: string;
-}
+  extends coreClient.OperationOptions {}
 
 /** Contains response data for the listByResourceGroupNext operation. */
 export type WorkflowListByResourceGroupNextResponse = WorkflowListResult;
@@ -513,6 +647,8 @@ export interface DeveloperHubServiceClientOptionalParams
   $host?: string;
   /** Api Version */
   apiVersion?: string;
+  /** The ManagedCluster resource associated with the workflows. */
+  managedClusterResource?: string;
   /** Overrides client endpoint. */
   endpoint?: string;
 }