+++ sdk/devhub/arm-devhub/src/models/parameters.ts
@@ -44,7 +44,7 @@ export const $host: OperationURLParameter = {
 export const apiVersion: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
-    defaultValue: "2022-04-01-preview",
+    defaultValue: "2022-10-11-preview",
     isConstant: true,
     serializedName: "api-version",
     type: {
@@ -136,7 +136,7 @@ export const resourceGroupName: OperationURLParameter = {
 };
 
 export const managedClusterResource: OperationQueryParameter = {
-  parameterPath: ["options", "managedClusterResource"],
+  parameterPath: "managedClusterResource",
   mapper: {
     serializedName: "managedClusterResource",
     type: {